CURRENT CONTROL FOR ALL LOAD WAWE FORMS
Our previous designs for "mosfet gate voltage control with opamp according to current measurement" was not doing well with coils and others waweforms rather than pure DC. "In this commit I have added A file FINISHED DUMMY LOAD MOSFET PART". Which works well for all waweforms and load characteristics and works perfect for pure DC. The  circuit tries to find the suitable gate voltage  at peak load voltage. Thus limits the current. However cant control the current only limits it  (can control it if it is ~pure DC).  Also added a snubber circuit to protect the mosfet. When a voltage spike occurs, the energy is transferred to a capacitor which is in paralel with a 92 OHM resistor. When capacitor is 48V the resistor can dissipate ~25Watts of power continuously. So if it reaches to that point, arduino shuts the relays and mosfets until the voltage across capacitor decreases to 10V and manuel toggle is required.