REVERSE & LOW VOLTAGE PROTECTION
floating votlage , Low voltage or negative voltage closese the  relay  and another opamp circuit holds it closed in 2ms electronical + Mechanical time until either
1) User manually preses to the button
2)Arduino shuts the safety( not obvious to all people who uses the device only to ones who read the manual)

Also arduino may deliberately close the relay. (over voltage or some preparing)

# ? # Why arduino may remove the protection ?
Say we have a square wave: in such case the if frequency is increased opamp's output voltage decreases(bandwith) and triggers the safety feature {LOW DUTY PWM may  also trigger } which turns off the relay. Arduino should anlyze the signal and if it decides that this is a square wawe, the safety feature may be removed. But if the wawe is sine and arduino thinks oh maybe I may remove the safety feature, at negative period of the wawe the reverse current may reach up to 500A 0ver mosfet's body diode; which is a catastrophe